// src/app/Model/StaffDetail.ts
export class StaffDetail {
    staffName: string = "";
    staffF_SName: string = "";
    staffDob: string = "";
    staffDateOfAppointment: string = "";
    staffDateOfConfirm: string = "";
    staffEduQual: string = "";
    staffProQual: string = "";
    staffDesignation: string = "";
    staffClassTaught: string = "";
    staffSubjentTaught: string = "";
    staffSalary: string = "";
}